created module for api; got and rendered data from server
