+++ js/main.js
@@ -1,8 +1,15 @@
 import './modules/page-state.js';
 import './modules/popup.js';
 import {addValidationForRooms, addValidationForMinPrice} from './modules/form.js';
-import './modules/map.js';
+import {renderPopups} from './modules/map.js';
+import {getData} from './modules/api.js';
 
 addValidationForRooms();
 addValidationForMinPrice();
+
+getData()
+  .then((data) => {
+    renderPopups(data);
+  });
+
