+++ js/modules/api.js
@@ -1,0 +1,8 @@
+const getData = () => fetch('https://23.javascript.pages.academy/keksobooking/data')
+  .then((response) => {
+    if (response.ok) {
+      return response.json();
+    }
+  });
+
+export {getData};
